@@ -9,36 +9,35 @@
   },
   "license": "LGPL-3.0-only",
   "private": false,
+  "main": "./build/lib/api.js",
   "files": [
     "contracts/*.sol",
-    "build/extracted/*.json"
+    "build/extracted/",
+    "build/lib/"
   ],
   "engines": {
     "yarn": "^1.19.2",
     "node": "^12"
   },
   "scripts": {
-    "lint:sol": "solhint 'contracts/**/*.sol'",
-    "format:sol": "prettier --write 'contracts/**/*.sol'",
-    "build:tsc": "rm -f migrations/*.js* && tsc",
-    "build:types": "typechain --target truffle --outDir './types/truffle-contracts' './build/contracts/*.json' && scripts/fix.sh",
-    "build:sol": "yarn lint:sol && truffle compile && yarn build:types",
-    "build": "yarn build:sol && yarn build:tsc",
-    "extract": "rm -rf 'build/extracted' && ts-node ./scripts/extract-build.ts",
-    "migrate": "yarn build && truffle migrate",
-    "test": "yarn build && scripts/test.sh",
-    "coverage": "yarn build && scripts/coverage.sh",
-    "network": "ONLY_NETWORK=true scripts/test.sh"
+    "build": "ts-node ./scripts/cli.ts build",
+    "coverage": "ts-node ./scripts/cli.ts coverage",
+    "fund": "ts-node ./scripts/cli.ts fund",
+    "migrate": "ts-node ./scripts/cli.ts migrate",
+    "network": "ts-node ./scripts/cli.ts network",
+    "test": "ts-node ./scripts/cli.ts test"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "npm run format:sol && yarn build && yarn extract && git add ."
+      "pre-commit": "yarn build && git add ."
     }
   },
   "dependencies": {
+    "@hoprnet/hopr-demo-seeds": "https://github.com/hoprnet/hopr-demo-seeds.git",
     "@hoprnet/hopr-utils": "https://github.com/hoprnet/hopr-utils.git",
     "@openzeppelin/contracts": "^2.4.0",
     "bignumber.js": "^9.0.0",
+    "ganache-core": "^2.10.2",
     "web3": "^1.2.4"
   },
   "devDependencies": {
@@ -48,19 +47,17 @@
     "@types/mocha": "^5.2.7",
     "@types/node": "^12.12.17",
     "chai": "^4.2.0",
-    "ganache-cli": "^6.7.0",
     "ganache-cli-coverage": "https://github.com/frangio/ganache-cli/releases/download/v6.4.1-coverage/ganache-cli-coverage-6.4.1.tgz",
     "husky": "^4.2.3",
     "mocha": "^6.2.2",
     "prettier": "^1.19.1",
-    "prettier-plugin-solidity": "^1.0.0-alpha.39",
     "solhint": "^2.3.0",
-    "solidity-coverage": "^0.6.7",
-    "truffle": "^5.1.3",
+    "solidity-coverage": "^0.7.0",
+    "truffle": "^5.1.22",
     "truffle-typings": "^1.0.8",
-    "ts-node": "^8.5.4",
+    "ts-node": "^8.9.0",
     "typechain": "^1.0.3",
     "typechain-target-truffle": "^1.0.2",
-    "typescript": "^3.7.3"
+    "typescript": "^3.8.3"
   }
 }
