@@ -1,8 +1,13 @@
+const { NODE_SEEDS, BOOTSTRAP_SEEDS } = require("@hoprnet/hopr-demo-seeds");
+
+const accounts = NODE_SEEDS.concat(BOOTSTRAP_SEEDS);
+const balance = Number(1000000000000000000000000).toString(16);
+
 module.exports = {
-  norpc: true,
-  testCommand:
-    "node --max-old-space-size=4096 ../node_modules/.bin/truffle test --network coverage",
-  compileCommand:
-    "node --max-old-space-size=4096 ../node_modules/.bin/truffle compile --network coverage",
-  copyPackages: ["@openzeppelin/contracts", "@openzeppelin/test-helpers"]
+  providerOptions: {
+    accounts: accounts.map(account => ({
+      secretKey: account,
+      balance
+    }))
+  }
 };
