@@ -1,5 +1,5 @@
 const HoprToken = artifacts.require("HoprToken");
 
-module.exports = (async deployer => {
+module.exports = async deployer => {
   await deployer.deploy(HoprToken);
-}) as Truffle.Migration;
+};
