@@ -1,42 +1,34 @@
-import { keccak256, signMessage } from "./random";
+import { keccak256, signMessage } from './random'
 
 type IFund = (args: {
-  web3: any;
-  stateCounter: string;
-  initiator: string;
-  deposit: string;
-  partyAAmount: string;
-  notAfter: string;
-  signerPrivKey: string;
+  web3: any
+  stateCounter: string
+  initiator: string
+  deposit: string
+  partyAAmount: string
+  notAfter: string
+  signerPrivKey: string
 }) => {
-  hashedFund: string; // return hashed alternative
-  signature: string; // signature of hashedTicket
-  r: string;
-  s: string;
-  v: string;
-};
+  hashedFund: string // return hashed alternative
+  signature: string // signature of hashedTicket
+  r: string
+  s: string
+  v: string
+}
 
 /*
   prepares fund payload
 */
-const Fund: IFund = ({
-  web3,
-  stateCounter,
-  initiator,
-  deposit,
-  partyAAmount,
-  notAfter,
-  signerPrivKey
-}) => {
+const Fund: IFund = ({ web3, stateCounter, initiator, deposit, partyAAmount, notAfter, signerPrivKey }) => {
   const hashedFund = keccak256(
-    { type: "uint256", value: stateCounter },
-    { type: "address", value: initiator },
-    { type: "uint256", value: deposit },
-    { type: "uint256", value: partyAAmount },
-    { type: "uint256", value: notAfter }
-  );
+    { type: 'uint256', value: stateCounter },
+    { type: 'address', value: initiator },
+    { type: 'uint256', value: deposit },
+    { type: 'uint256', value: partyAAmount },
+    { type: 'uint256', value: notAfter }
+  )
 
-  const { signature, r, s, v } = signMessage(web3, hashedFund, signerPrivKey);
+  const { signature, r, s, v } = signMessage(web3, hashedFund, signerPrivKey)
 
   return {
     hashedFund,
@@ -44,7 +36,7 @@ const Fund: IFund = ({
     r,
     s,
     v
-  };
-};
+  }
+}
 
-export { Fund };
+export { Fund }
