@@ -1,38 +1,38 @@
-import BigNumber from "bignumber.js";
-import { keccak256, xorBytes32, signMessage, MAX_UINT256 } from "./random";
+import BigNumber from 'bignumber.js'
+import { keccak256, xorBytes32, signMessage, MAX_UINT256 } from './random'
 
-BigNumber.config({ EXPONENTIAL_AT: 1e9 });
+BigNumber.config({ EXPONENTIAL_AT: 1e9 })
 
 type ITicket = (args: {
-  web3: any;
-  accountA: string;
-  accountB: string;
-  porSecretA: string; // needs to be bytes32
-  porSecretB: string; // needs to be bytes32
-  signerPrivKey: string;
-  counterPartySecret: string; // needs to be bytes32
-  amount: string;
-  counter: number;
-  winProbPercent: string; // max 100
+  web3: any
+  accountA: string
+  accountB: string
+  porSecretA: string // needs to be bytes32
+  porSecretB: string // needs to be bytes32
+  signerPrivKey: string
+  counterPartySecret: string // needs to be bytes32
+  amount: string
+  counter: number
+  winProbPercent: string // max 100
 }) => {
-  accountA: string; // return same as provided
-  accountB: string; // return same as provided
-  porSecretA: string; // return same as provided
-  porSecretB: string; // return same as provided
-  counterPartySecret: string; // return same as provided
-  amount: string; // return same as provided
-  counter: number; // return same as provided
-  hashedPorSecretA: string; // return hashed alternative
-  hashedPorSecretB: string; // return hashed alternative
-  challenge: string; // return hashed alternative
-  hashedCounterPartySecret: string; // return hashed alternative
-  winProb: string; // return winProb in bytes32
-  hashedTicket: string; // return hashed alternative
-  signature: string; // signature of hashedTicket
-  r: string;
-  s: string;
-  v: string;
-};
+  accountA: string // return same as provided
+  accountB: string // return same as provided
+  porSecretA: string // return same as provided
+  porSecretB: string // return same as provided
+  counterPartySecret: string // return same as provided
+  amount: string // return same as provided
+  counter: number // return same as provided
+  hashedPorSecretA: string // return hashed alternative
+  hashedPorSecretB: string // return hashed alternative
+  challenge: string // return hashed alternative
+  hashedCounterPartySecret: string // return hashed alternative
+  winProb: string // return winProb in bytes32
+  hashedTicket: string // return hashed alternative
+  signature: string // signature of hashedTicket
+  r: string
+  s: string
+  v: string
+}
 
 /*
   prepares ticket payload
@@ -50,15 +50,15 @@ const Ticket: ITicket = ({
   winProbPercent
 }) => {
   // proof of relay related hashes
-  const hashedPorSecretA = keccak256({ type: "bytes32", value: porSecretA });
-  const hashedPorSecretB = keccak256({ type: "bytes32", value: porSecretB });
-  const challenge = xorBytes32(hashedPorSecretA, hashedPorSecretB);
+  const hashedPorSecretA = keccak256({ type: 'bytes32', value: porSecretA })
+  const hashedPorSecretB = keccak256({ type: 'bytes32', value: porSecretB })
+  const challenge = xorBytes32(hashedPorSecretA, hashedPorSecretB)
 
   // proof of randomness related hashes
   const hashedCounterPartySecret = keccak256({
-    type: "bytes32",
+    type: 'bytes32',
     value: counterPartySecret
-  });
+  })
 
   // calculate win probability in bytes32
   const winProb = web3.utils.numberToHex(
@@ -66,17 +66,17 @@ const Ticket: ITicket = ({
       .multipliedBy(MAX_UINT256)
       .dividedBy(100)
       .toString()
-  );
+  )
 
   const hashedTicket = keccak256(
-    { type: "bytes32", value: challenge },
-    { type: "bytes32", value: counterPartySecret },
-    { type: "uint256", value: counter },
-    { type: "uint256", value: amount },
-    { type: "bytes32", value: winProb }
-  );
+    { type: 'bytes32', value: challenge },
+    { type: 'bytes32', value: counterPartySecret },
+    { type: 'uint256', value: counter },
+    { type: 'uint256', value: amount },
+    { type: 'bytes32', value: winProb }
+  )
 
-  const { signature, r, s, v } = signMessage(web3, hashedTicket, signerPrivKey);
+  const { signature, r, s, v } = signMessage(web3, hashedTicket, signerPrivKey)
 
   return {
     accountA,
@@ -96,7 +96,7 @@ const Ticket: ITicket = ({
     r,
     s,
     v
-  };
-};
+  }
+}
 
-export { Ticket };
+export { Ticket }
